move Home to containers folder, fix login issue, update tests
